+++ test/test.ff4-properties.js
@@ -10,6 +10,7 @@
         var AP = properties.typeOf.AP;
 
         t.deepEqual( AP.parse(["myapp:1.2.3"]), ["myapp", "1.2.3"] );
+        t.deepEqual( AP.stringify(["myapp", "1.2.3"]), ["myapp:1.2.3"] );
 
         t.end();
     });
@@ -20,6 +21,9 @@
         t.equal( SZ.parse(["19"]), 19 );
         t.deepEqual( SZ.parse(["12:34"]), [12, 34] );
 
+        t.deepEqual( SZ.stringify(19), ["19"] );
+        t.deepEqual( SZ.stringify([12, 34]), ["12:34"] );
+
         t.end();
     });
 
@@ -28,6 +32,9 @@
 
         t.equal( FG.parse([""]), null );
         t.deepEqual( FG.parse(["257:fig"]), [257, "fig"] );
+
+        t.deepEqual( FG.stringify(null), [""] );
+        t.deepEqual( FG.stringify([257, "fig"]), ["257:fig"] );
 
         t.end();
     });